+++ evalml/utils/nullable_type_utils.py
@@ -1,4 +1,5 @@
 import pandas as pd
+import woodwork as ww
 from woodwork.logical_types import AgeNullable, BooleanNullable, IntegerNullable
 
 DOWNCAST_TYPE_DICT = {
@@ -57,7 +58,10 @@ def _downcast_nullable_y(y, handle_boolean_nullable=True, handle_integer_nullabl
         y with any incompatible nullable types downcasted to compatible equivalents.
     """
     if y.ww.schema is None:
-        y.ww.init()
+        if isinstance(y, pd.DataFrame):
+            y.ww.init()
+        else:
+            y = ww.init_series(y)
 
     incompatible_logical_types = _get_incompatible_nullable_types(
         handle_boolean_nullable,