+++ evalml/pipelines/utils.py
@@ -132,6 +132,7 @@ def _get_datetime(X, y, problem_type, estimator_class, sampler_name=None):
     if add_datetime_featurizer and estimator_class.model_family not in [
         ModelFamily.ARIMA,
         ModelFamily.PROPHET,
+        ModelFamily.VARMAX,
     ]:
         components.append(DateTimeFeaturizer)
     return components
@@ -298,13 +299,7 @@ def _get_preprocessing_components(
     Returns:
         list[Transformer]: A list of applicable preprocessing components to use with the estimator.
     """
-    if is_multiseries(problem_type):
-        if include_decomposer:
-            components_functions = [_get_decomposer]
-        else:
-            return []
-
-    elif is_time_series(problem_type):
+    if is_time_series(problem_type):
         components_functions = [
             _get_label_encoder,
             _get_drop_all_null,
@@ -1508,7 +1503,7 @@ def stack_X(X, series_id_name, time_index, starting_index=None, series_id_values
         time_index (str): The name of the time index column.
         starting_index (int): The starting index to use for the stacked DataFrame. If None, the starting index
             will match that of the input data. Defaults to None.
-        series_id_values (set, list): The unique values of a series ID, used to generate the index. If None, values will
+        series_id_values (list): The unique values of a series ID, used to generate the index. If None, values will
             be generated from X column values. Required if X only has time index values and no exogenous values.
             Defaults to None.
 
@@ -1516,14 +1511,21 @@ def stack_X(X, series_id_name, time_index, starting_index=None, series_id_values
         pd.DataFrame: The restacked features.
     """
     original_columns = set()
-    series_ids = series_id_values or set()
-    if series_id_values is None:
+    if series_id_values is not None:
+        series_ids = series_id_values
+    else:
+        # Using list to maintain order (vs. a set)
+        series_ids = list()
         for col in X.columns:
             if col == time_index:
                 continue
             separated_name = col.split(MULTISERIES_SEPARATOR_SYMBOL)
             original_columns.add(MULTISERIES_SEPARATOR_SYMBOL.join(separated_name[:-1]))
-            series_ids.add(separated_name[-1])
+            series_ids.append(separated_name[-1])
+        # Remove duplicates while maintaining insertion order
+        # Need order to match series ID labels correctly when restacking columns
+        seen = set()
+        series_ids = [val for val in series_ids if not (val in seen or seen.add(val))]
 
     if len(series_ids) == 0:
         raise ValueError(
@@ -1542,7 +1544,7 @@ def stack_X(X, series_id_name, time_index, starting_index=None, series_id_values
         restacked_X = pd.DataFrame(
             {
                 time_index: time_index_col,
-                series_id_name: sorted(list(series_ids)) * len(X),
+                series_id_name: list(series_ids) * len(X),
             },
             index=stacked_index,
         )