+++ evalml/tests/automl_tests/test_iterative_algorithm.py
@@ -18,7 +18,6 @@ from evalml.pipelines.components import (
     DateTimeFeaturizer,
     EmailFeaturizer,
     NaturalLanguageFeaturizer,
-    STLDecomposer,
     TimeSeriesFeaturizer,
     URLFeaturizer,
 )
@@ -98,7 +97,6 @@ def test_iterative_algorithm_init(
     assert algo.batch_number == 0
     assert algo.default_max_batches == 1
     estimators = get_estimators(problem_type)
-    decomposer = [STLDecomposer] if is_regression(problem_type) else []
     assert len(algo.allowed_pipelines) == len(
         [
             make_pipeline(
@@ -106,11 +104,15 @@ def test_iterative_algorithm_init(
                 y,
                 estimator,
                 problem_type,
+                include_decomposer=include_decomposer,
                 parameters=search_parameters,
             )
             for estimator in estimators
-        ]
-        + decomposer,
+            # Generate both decomposer and non-decomposer pipelines when problem type is multiseries time series reg.
+            for include_decomposer in (
+                [True, False] if is_regression(problem_type) else [False]
+            )
+        ],
     )