+++ docs/source/release_notes.rst
@@ -4,6 +4,7 @@ Release Notes
     * Enhancements
     * Fixes
         * Fixed bug in `_downcast_nullable_y` causing woodwork initialization issues :pr:`4369`
+        * Fixed multiseries prediction interval labels :pr:`4377`
     * Changes
         * Pinned scipy version to under 1.12.0 :pr:`4380`
     * Documentation Changes